Add Calico IPAM plugin to manage pod IP addresses
In the configuration and managment of the network with Calico it needs a
mechanism of doing IP address management (commonly called IPAM).  At the
present time we do not ship the Calico IPAM plugin which increases the
complexity of the CNI configuration due to more manual steps needing to
be taken by the end user.