Add annotation for sidecar ordering
This commits adds sidecar containers support using a pod annotation. The
format is:
	alpha.kinvolk.io/sidecar: '["container-1", "container-2"]'

Note that the value of the annotation needs to be valid JSON, so each
container name needs to be enclosed with double quotes.

If the annotation is not present or malformed, the behavior is as this
patch was not present: no contianer is considered a sidecar.
Non-existing containers referenced in the annotation are ignored.