Relax restriction for usernamespace with Pod mode
There are some features that *could* be not compatible with user namespaces:
- hostpath volumes
- PVC
- sharing other host namespaces
- some capabilities

The current code forbids to use 'Pod' mode when one of those features is present
on the PodSpec. This commit relaxes that logic and replaces it by a warning only
for the case where host namespaces are shared given that it's the only case
that could present issues when creating the container.