WIP: add prints to debug no API server case
I tested this:

	1. start kubernetes with: ./hack/local-up-cluster.sh
	2. Create the pod added in previous commit: sidecar-tests/noprestop-test.yaml.
	Check the previous commit msg for instructions on how to do that (is not just apply)
	3. See containers are started in order by looking at the pod logs
	generated in /tmp/logs/log.txt
	4. Kill the API server (ps and kill the process), see kubelet works
	fine by looking at pod logs in /tmp/logs/log.txt
	5. Run docker ps and kill both containers in the pod with:
		docker stop <container id>
	6. See kubelet logs:
		tail -F /tmp/kubelet.log | grep XXX

	It says all containers are ready
	7. See /tmp/logs/log.txt, see readiness probe say the sidecar is not
	yet ready, but non-sidecar is started anyways
	8. See in the pod logs that containers are started without waiting for
	sidecar to be ready

This indeed violates the guarantees for sidecar containers (not in this patch
exactly, as we don't restart sidecars in order in this patch). However,
this patch shows that the way we are checking for containers ready
status is not reliable if the API server is down.

It is still left to investigate other ways to check for the readiness (the
readiness probes are executed by the kubelet, so there might be other
ways).

If no other ways to check containers readiness status are found, then
more plumbing is needed to properly support complex sidecar containers
with restart policies that are not just restarting individual
containers.