WIP: test race condition on current patches
The current code uses the pod status cached from the API server to retrieve
containers ready status. When the API server is down, this is of course
not updated and therefore the following situation can happen:
 * A pod has been scheduled to a node
 * The API server goes down before the pod sidecars are ready
 * Then the non-sidecar containers of the pod will not be started until the API server is back

That is because we rely on the podstatus from the apiserver to know when
a container is ready and only start non-sidecars after sidecars are
ready. Instead, something similar as done for livenessProbes can be used
to circumvent this edge-case.

However, once all containers have been started the ready state of
containers is not used to restart them, so once all have started once,
they work fine.

This race condition can be observed with the following steps:
 1. Start kubernetes
 2. Check the previous commit msg for instructions on how to setup the
    environment for the tests in sidecar-tests/
 3. Make sure the the log.sh script you are using has a "sleep 60"
    instead of "sleep 10" (i.e. you are not using a stale script)
 4. Apply the apiserver-down.yaml added in this commit.
 5. Run "ps faux | grep api" and kill the api server pid
 6. See in /tmp/logs/log.txt that only the sidecar has been started and
    it eventually the sidecar container is ready (after 1 minute).
    Please note the other container in the pod is not started even after
    these log show that it is ready.
 7. Run "tail -F /tmp/kubelet.log | grep XXX" to see the added debug
    logs. You will see something like:

	I0106 12:49:37.791398   76579 prober.go:166] XXX rata: r == probe.Success:true, s: , e: <nil>
	I0106 12:49:38.418355   76579 kuberuntime_manager.go:507] XXX rata: Status is not ready for pod: "prestop-6db4764597-4z89c", container: "sidecar-1"
	I0106 12:49:38.418400   76579 kuberuntime_manager.go:507] XXX rata: Status is not ready for pod: "prestop-6db4764597-4z89c", container: "debian-1"

The first line means that the readiness probe was executed and the result is
equal to probe.Success (i.e. it returns true).

The second and third line show the status as seen by the kubelet
component that starts the container, which shows both container are seen
as not ready.

Therefore, the unintended behavior is hit under this condition.

Please note also that if you start the apiserver again (just do a c&p
from the output of the "ps" line), the pod startup continues just fine.